requirements.txt
Flask==3.0.0
fpdf2==2.7.5
gunicorn==21.2.0
matplotlib==3.8.0
pandas==2.1.1
psycopg2-binary==2.9.9
scikit-learn==1.3.1
seaborn==0.13.0
SQLAlchemy==2.0.21